address PR comments with multiple changes + move adv filtering logic to a mixin + add missing translations
